feat: refactor Electron app structure and update main window handling
