+++ src/electron/app.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@mariodebono/di';
+import { ElectronApp } from './electron-app.js';
+
+@Module({
+    providers: [ElectronApp],
+})
+export class AppModule {}
